assignment/requirements.txt
Django
djangorestframework
django-filter
django-rest-knox
pyyaml
uritemplate
django-rest-knox
